Break out the collision logic into an explicit collider
This frees the physics engine from knowing what a "solid" is. You
just pass it a function that accepts a tile index and returns whether
that should count as a collision, and that function takes care of any
special logic. We'll want a few of these to handle variable solidity
for glitch states especially.