Add a title theme, queue it up as the first level theme
This is global and slightly hacky: every exit door now queues up
the normal level music, which means the *second* level will replace
the title theme with the proper music. The title theme has no
variants, so it ignores dead and hot zones; feel free to use these
for mood on the title level.